Aux in umap. Still need to clean
